+++ pywb/recorder/test/test_recorder.py
@@ -6,8 +6,8 @@ from pywb.warcserver.test.testutils import FakeRedisTests
 
 import os
 import webtest
+import pytest
 
-from pytest import raises
 from fakeredis import FakeStrictRedis
 
 from pywb.recorder.recorderapp import RecorderApp
@@ -607,6 +607,7 @@ class TestRecorder(LiveServerTests, FakeRedisTests, TempDirTests, BaseTestClass)
         assert len(writer.fh_cache) == 0
 
     def test_record_video_metadata(self):
+        pytest.importorskip('youtube_dl')
         warc_path = to_path(self.root_dir + '/warcs/{user}/{coll}/')
 
         dedup_index = self._get_dedup_index()