+++ pywb/warcserver/test/test_handlers.py
@@ -7,6 +7,7 @@ from six.moves.urllib.parse import urlencode
 import webtest
 from fakeredis import FakeStrictRedis
 from mock import patch
+import pytest
 
 import json
 
@@ -380,6 +381,7 @@ foo=bar&test=abc"""
         assert resp.headers['Memento-Datetime'] == 'Mon, 29 Jul 2013 19:51:51 GMT'
 
     def test_live_video_loader(self):
+        pytest.importorskip('youtube_dl')
         params = {'url': 'http://www.youtube.com/v/BfBgWtAIbRc',
                   'content_type': 'application/vnd.youtube-dl_formats+json'
                  }
@@ -397,6 +399,7 @@ foo=bar&test=abc"""
         assert b'Content-Type: application/vnd.youtube-dl_formats+json' in resp.body
 
     def test_live_video_loader_post(self):
+        pytest.importorskip('youtube_dl')
         req_data = """\
 GET /v/BfBgWtAIbRc HTTP/1.1
 accept-encoding: gzip, deflate